Fixes crashes when using numpy log on a batch transform data panel.
The recent change to the creation of the data panel ended up with
a panel with the dtype of 'object', which was causing numpy ufuncs
like `log` to crash out on an `AttributeError`.

This forces all frames in the panel to use a dtype of 'float',
we may want to look at seeting a dtype on a frame by frame basis,
e.g. 'volume' may more accurately be 'int'.

+++ zipline/transforms/utils.py
@@ -472,7 +472,8 @@ class BatchTransform(EventWindow):
         # user-overloaded get_value() method.
         data_dict = {tick['dt']: tick['data'] for tick in self.ticks}
         data = pd.Panel(data_dict, major_axis=self.field_names,
-                        minor_axis=self.sids)
+                        minor_axis=self.sids,
+                        dtype='float')
 
         if self.supplemental_data:
             # item will be a date stamp
